+++ cmd/broker/broker.go
@@ -10,7 +10,9 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/Azure/open-service-broker-azure/pkg/api/filters"
 	"github.com/Azure/open-service-broker-azure/pkg/api/filters/authenticator/basic" //nolint: lll
+	"github.com/Azure/open-service-broker-azure/pkg/api/filters/headers"
 	"github.com/Azure/open-service-broker-azure/pkg/broker"
 	"github.com/Azure/open-service-broker-azure/pkg/crypto/aes256"
 	"github.com/Azure/open-service-broker-azure/pkg/version"
@@ -114,12 +116,19 @@ func main() {
 		log.Fatal(err)
 	}
 
+	filterChain := filters.NewFilterChain(
+		[]filters.Filter{
+			authenticator,
+			headers.NewValidator(),
+		},
+	)
+
 	// Create broker
 	broker, err := broker.NewBroker(
 		storageRedisClient,
 		asyncRedisClient,
 		codec,
-		authenticator,
+		filterChain,
 		modules,
 		modulesConfig.MinStability,
 		azureConfig.DefaultLocation,
