+++ pkg/api/filters/authenticator/always/authenticator.go
@@ -6,17 +6,17 @@ import (
 	"github.com/Azure/open-service-broker-azure/pkg/api/filters"
 )
 
-// alwaysAuthenticator is a implementation of the filters.Filter
+// alwaysAuthenticator is a implementation of the filter.Filter
 // interface useful for testing. It unconditionally authenticates all requests.
 type alwaysAuthenticator struct{}
 
-// NewAuthenticator returns an implementation of the filters.Filter
+// NewAuthenticator returns an implementation of the filter.Filter
 // interface useful for testing. It unconditionally authenticates all requests.
 func NewAuthenticator() filters.Filter {
 	return &alwaysAuthenticator{}
 }
 
-func (a *alwaysAuthenticator) Execute(
+func (a *alwaysAuthenticator) Filter(
 	handler http.HandlerFunc,
 ) http.HandlerFunc {
 	return handler
